pyproject.toml
[project]
name = "keyword-spotting"
version = "0.1.0"
description = "Wake up your device with a keyword!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cython>=3.1.1",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "pytest>=8.4.0",
    "python-speech-features>=0.6",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "torch>=2.7.1",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
log_cli = 1

[tool.ruff]
select = [
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "E",  # error
  "F",  # pyflakes
  "I",  # isort
]
ignore = [
  "E501", # line too long
  "E741", # ambiguous variable name
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

# Same as Black.
line-length = 88

# Assume Python 3.8
target-version = "py38"

[tool.ruff.extend-per-file-ignores]
"__init__.py" = [
  "F401", # imported but unused
]
